Move icons folder
