+++ assets/blocks/lesson-actions/lesson-actions-block/index.js
@@ -9,7 +9,7 @@ import { __ } from '@wordpress/i18n';
 import metadata from './block.json';
 import edit from './lesson-actions-edit';
 import save from './lesson-actions-save';
-import icon from '../../../images/svg-icons/buttons.svg';
+import icon from '../../../icons/buttons.svg';
 
 export default {
 	title: __( 'Lesson Actions', 'sensei-lms' ),
