+++ assets/blocks/lesson-properties/index.js
@@ -6,7 +6,7 @@ import { __ } from '@wordpress/i18n';
 /**
  * Internal dependencies
  */
-import icon from '../../images/svg-icons/lesson-properties.svg';
+import icon from '../../icons/lesson-properties.svg';
 import metadata from './block.json';
 import edit from './lesson-properties-edit';
