+++ assets/extensions/extension-actions.js
@@ -10,7 +10,7 @@ import { check } from '@wordpress/icons';
  * Internal dependencies
  */
 import { EXTENSIONS_STORE, isLoadingStatus } from './store';
-import UpdateIcon from '../images/svg-icons/update.svg';
+import UpdateIcon from '../icons/update.svg';
 import { logEvent } from '../shared/helpers/log-event';
 import { getWoocommerceComPurchaseUrl } from '../shared/helpers/woocommerce-com';
